feat: add support to use PATs instead of OAuth (#361)
feat: add support for using tokens instead of OAuth

Tools will now be able to specify, in their metadata, that their OAuth
credential supports personal access tokens with the supportsOAuthTokenPrompt
field being "true". If a tool does this, then the OAuth credential will
prompt the user for all tokens listed in prompt_tokens in the credential
as sensitive and prompt_vars as not sensitive.

This change includes this field on the GitHub, Atlassian, and Slack tools.

Lastly, this change removes support for OAuth outside Obot. The OAuth
credential tool will no longer look at the CLI config nor default to use
the Gateway.

Signed-off-by: Donnie Adams <donnie@acorn.io>