enhance: updates to prompting framework (#404)
The prompting framework in GPTScript has been updated to support
field-level sensitivity and descriptions. This change bumps the Go
GPTScript dependency for these updates.

Additionally, the credentials here are updated to support more granular
and general configuration.

Finally, a validation tool is added to both Obot credentials. Validation is added for Bluesky (generic) and GitHub (OAuth) as
examples.

Signed-off-by: Donnie Adams <donnie@acorn.io>