feat: add new virus total provider for file scanning (#586)
feat: add file scanner provider tools for virus scanning

This change adds two such tools: Virus Total and MetaDefender.
These tools work very similarly to the model provider tools we
have in that they are configured, deconfigured, validated, started,
and stopped in the same manor.

Both file scanner tools are daemon tools that provide a files endpoint
that sends the file to the provider for virus scanning. If the file
is found to be malicious, then a 403 response is returned by the tool.

Signed-off-by: Donnie Adams <donnie@acorn.io>